requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-countries==6.1.3
django-heroku==0.3.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
Pillow==7.2.0
psycopg2==2.8.6
PyJWT==1.7.1
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.1
reportlab==3.5.55
requests==2.25.0
six==1.15.0
sqlparse==0.3.1
stripe==2.55.0
twilio==6.47.0
urllib3==1.26.2
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
